Add IFocusMenuExtension extension
This will allow plugins to express that a menu should not take focus
away from a player without the player doing something to enable that focus.

Also included is the ability to query its focus state, for whether you
want to time a menu out in the case it has never been focused or such.

Co-authored-by: Dawid "daffyyyy" Bepierszcz <41084667+daffyyyy@users.noreply.github.com>